Temporarily preserve parameter metadata when present
Summary:
Running `javac` 8 or above with `-parameters` includes some metadata about parameter names
in the resulting `.class` files. This metadata is intended for use at runtime by Reflection,
and thus is not technically part of the ABI. However, Buck's class ABI generation was not
stripping them out.

This would be fine (most people don't run with `-parameters` anyway), however, when the metadata
is present, the compiler will helpfully use it when constructing `Element`s to give to
annotation processors. Some annotation processors (Litho) have taken a dependency on this behavior,
so for now we must bring source ABIs to parity. I intend to work with Litho to remove the
dependency and then start stripping this metadata out of the ABIs.

Test Plan: New unit test

Reviewed By: dreiss

fbshipit-source-id: 9f08757