Coalesce a huge number of buck-out/gen and buck-out/annotation <component/> tags to just two
Summary:
This addresses a report of “fairly consistent lag while ... typing” which I hadn’t seen in files that didn’t directly refer to generated code but which I did see in a file that does. This diff seems to make that lag go away.

Secondarily, it saves **a lot** of XML and (presumably) a corresponding amount of Java heap. The lag-reduction is presumably due more to less data to traverse than to less memory pressure ... but every little bit helps.

Test Plan: Generate or regenerate a Project View; open it up. Works as before, with less lag in files that refer to generated code. Reporting SWE says "so far I do think it seems faster!"

Reviewed By: mkillianey

fbshipit-source-id: 36669a1