source_only_abi_deps
Summary:
We already have `required_for_source_only_abi`, which ensures that a given rule is always
available during source-only ABI gen for *all* of its dependents. That's appropriate for
things like annotations or enums where literally every rule that depends upon it will need it.

However, some cases are specific to exactly what the dependent is doing. For these cases,
we allow each rule to specify which of its dependencies are needed for correct source-only ABI
generation.

Test Plan: CI for regressions

Reviewed By: asp2insp

fbshipit-source-id: 6472908