Use StringTemplate instead of ClosureTemplate for buck internal webserver
Summary:
Closure compiler jar is 1.3 MB, and its dependency, icu4j, is 11 MB.

The templates were not particularly complex, and we are already using ST in
other places of the code, so it was easy to switch. `buck.pex` goes from 94MB
to 83MB after this change.

Also clean up dependencies in httpserver's buck file....

Test Plan: look at the 3 pages that's generated by the webserver, see tha they look the same.

Reviewed By: styurin

fbshipit-source-id: 0aa69de