Improve ErrorLogger
Summary:
This moves the logic of showing stack traces for internal errors into
the ErrorLogger and makes several improvements to the output.

1. When suppressing the stack trace, include the exception's class name
in the output. For many exceptions, just showing the message is really
confusing.
2. When including the stack trace, filter out frames that are included
in the WrapsException's stack trace. This means that to a user, the
context replaces that part of the stack trace. See the next diff for a
good comparison.

Test Plan: CI, updated tests

Reviewed By: ttsugriy

fbshipit-source-id: e456b02