Fix handling of temporary artifact zip
Summary:
Currently, if the temporary artifact zip is borrowed (i.e. moved) by the
cache's store() call, the ArtifactUploader will still try to delete it
in a listener on the future. This then throws a NoSuchFileException
which gets printed to the user (but doesn't cause failures).

This changes ArtifactUploader to only try to delete the file if it
hasn't been borrowed.

Test Plan: CI, updated InMemoryArtifactStore to borrow the path if it's borrowable

Reviewed By: ttsugriy

fbshipit-source-id: 051f6df