Fix ApkGenrule's runtime deps
Summary:
Any implementation of HasInstallableApk should have runtime deps on
any rule that creates one of the files exposed in the rule's ApkInfo.

This adds HasInstallableApk.getRuntimeDepsForInstallableApk() that computes
this for any implementors.

It updates both ApkGenrule and AndroidBinary to use that for computing their
runtime deps.

We then no longer need some weird hack in AndroidInstrumentationTest to dig in
to find the underlying apk (which was done to get the manifest, which is really
quite brittle and didn't really fix the case for doing `buck install
some_apk_genrule`).

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 1722f82