Add a rule key diff printer to consistency checker
Summary: This adds a class that will print out differences between rule keys while traversing a graph of rulekeys. The actual logic to traverse that graph will be coming in a subsequent diff.

Test Plan: Added unit tests. Also used it locally with a sample project to see that it printed out reasonably.

Reviewed By: ttsugriy

fbshipit-source-id: 985670a