Exclude javac-plugin.jar from buck hash in ant build
Summary:
The generated jar contains the Java version including the minor version and that's causing the generated hash to be diverge when using different JDKs.

Solution is to remove `javac-plugin.jar` from the list of files used to calculate the hash in ant build.

Test Plan: Build buck with ant on different JDKs, check that hash is the same.

Reviewed By: ttsugriy

fbshipit-source-id: 75f1f14