Split the metadata directory into three subdirs
Summary:
This adds "artifact", "build", and "other" subdirectories to the
metadata directory. Previously we just mixed all kinds of random
metadata together and once it was written to disk we could no longer
really tell what type of metadata it was meant to be.

Now, metadata from the BuildInfoStore (added with addBuildMetadata()) is put
in the "build" directory (or in the sqlite db when using the
SqliteBuildInfoStore).

Metadata that is added directly into the artifact
(from addMetadata()) is put in the "artifact" directory.

The manifest key is put in the "other" directory. This is a weird piece
of metadata that isn't included in artifact uploads at all.

The on-disk state is required to be a source-of-truth for builds (a
new buckd can only read state from disk). Splitting it up so that we can
tell what type of metadata everything is will allow us to construct
artifacts for uploading from the on-disk state instead of the in-memory
state (which should be less error-prone).

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: a8d412b