Change invalidateInitializeFromDiskState() to outputsWillChange()
Summary:
Before building locally or fetching from cache we currently call
invalidateInitializeFromDiskState(). This renames that to
outputsWillChange() and adds a call
`onDiskBuildInfo.deleteExistingMetadata()` which will delete both
"build" and "artifact" metadata.

Removes the explicit deleting of "build" and "artifact" metadata before
upoading and just rely on outputsWillChange() to do that.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: 95797ac