Make artifact upload be based on the OnDiskBuildInfo
Summary:
Previously, artifact uploads were done based on a mix of the state from
the OnDiskBuildInfo and from the BuildInfoRecorder. This has in the past
caused several issues where those two get out of sync.

Importantly, in the fetched from cache flow, the OnDiskBuildInfo has to
be correctly updated or future builds will fail. In that case, if the
BuildInfoRecorder is not updated, future builds will be fine, but we'll
possibly create an invalid artifact that will then corrupt later builds.
The second case is much harder to detect/diagnose.

This switches artifact uploading to be based entirely on the
OnDiskBuildInfo.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: 0f80c47