Discard CachingBuildRuleBuilder.buildInfoRecorder after finalizing outputs
Summary:
This adds a simple Discardable class that just holds some instance of
some type and has a discard() that disables access to the instance. This
can be used basically the same as having a `Nullable` field, but it's a
little bit more friendly w/ infer and it's only one-way (i.e. the field
only goes from set to null).

Use this in CachingBuildRuleBuilder to discard the BuildInfoRecorder.
This is just a defensive measure to prevent use of that field as it may
not accurately reflect the on-disk state of the build rule's
outputs/metadata.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: 27a6c9a