Split up finalizeBuiltLocallyOrFetchedFromCache
Summary:
This is an odd function that has a bunch of conditional blocks that only
happen for one or the other of built locally/fetched from cache and it
is difficult to reason about.

This splits this into two functions `finalizeBuiltLocally()` and
`finalizeFetchedFromCache()` and simplifies the conditionals to remove
the cases that cannot happen in the respective functions.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: 30a17fe