Do not load information into the BuildInfoRecorder when fetching from cache
Summary:
Previously the fetch flow was something like:
1. write all metadata to the buildInfoStore
2. explicitly list some keys and load those into the buidInfoRecorder
3. do a bunch of other stuff
4. write some other metadata to the buildInfoRecorder
5. write everything in the buildInfoRecorder to the buildInfoStore

This is rather error-prone, especially (2) and the fact that 1/2/4+5 are
very much split from each other and hard to reason about. This changes the
`finalizeFetcheFromCache()` function to mostly depend on the
OnDiskBuildInfo instead of the BuildInfoRecorder (and to assert that
only expected keys have been recorded). This allows us to remove (2) and
to make (5) really just write the keys that we expect.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: 7e15e82