Create a separate resource file installer
Summary:
This adds a new rule for installing resource files separate from the full ExopackageFilesInstaller.
This allows us to install resource files as soon as they have been created.

We construct a single ExopackageResourcesInstaller for each creator of exopackage resources.
This is done because the installers will synchronize on the underlying AndroidDevicesHelper
and so we don't want a single rule to generate a bunch of resource files and then take up a
bunch of build threads all waiting on each other.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 123fed8