Create a separate AndroidManifest.xml copying rule
Summary:
For exopackage installation, we use the manifest to determine the
package name. Currently, this is acquired from a step copying it in
AndroidBinary. For concurrent install, so that we don't need to wait for the
AndroidBinary rule to install files, split this out into a separate build rule.

The copy of it in the AndroidBinary rule stays to be backwards compatible with
things outside of BUCK that would read the AndroidManifest.xml after building
an apk.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: a6985d8