Sanitize preprocessor flags
Summary:
When using the following argument in an apple_library and cxx_library.

```
    preprocessor_flags = ["-I$SDKROOT/usr/include/libxml2"],
```

It will create a rulekey inconsistency when the path of the SDK changes.

This diff sanitize this value when it's included in the rulekey.

Test Plan:
Unit tests + CI

```
buck targets [target] --show-rulekey --show-transitive-rulekeys
```

For a [target] with ` preprocessor_flags = ["-I$SDKROOT/usr/include/libxml2"]`.

Reviewed By: ttsugriy

fbshipit-source-id: e785438