Introduce AndroidLegacyToolchain
Summary:
Introducing `AndroidLegacyToolchain` that provides access to `AndroidDirectoryResolver` and `AndroidPlatformTarget` and using it to access `AndroidPlatformTarget` instead of accessing it through `ExecutionContext`.

The purpose of this change is to remove `AndroidPlatformTarget` from `ExecutionContext` to drop the `c/f/buck/step:step` dependency on Android.

The proper way would be to use `AndroidSdk` to provide access to necessary functionality, but it would require pretty massive refactoring to avoid circular dependencies. Using a temporary class during migration allows much easier transitioning to the new design.

The common rule for using `ToolchainProvider` is to get all necessary toolchains inside `Description` classes and pass these toolchains downstream to the steps and other classes. This will make it easier in future to add support for multiple toolchains with different names.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: 34bd0aa