Use AndroidLegacyToolchain to access AndroidPlatformTarget instead of ExecutionContext, part 3
Summary:
This change continues the replacement of `ExecutionContext` to access `AndroidPlatformTarget` with `AndroidLegacyToolchain`.

The common rule for using `ToolchainProvider` is to get all necessary toolchains inside `Description` classes and pass these toolchains downstream to the steps and other classes. This will make it easier in future to add support for multiple toolchains with different names.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: e7af016