BuildPhase now waits for all slaves to finish before returning
Summary: There was previously a bug where the local client would try to download BuildSlaveFinishedStats before they were available. This is because it only waiting for the top level BuildStatus in the BuildJob to be set to finished, but didn't wait for the individual slaves to set their status (which happens as the final step, after they have uploaded BuildSlaveFinishedStats).

Test Plan: ci

Reviewed By: shivanker

fbshipit-source-id: cfba190