Coordinator can now choose a random port for itself.
Summary:
We used to check for a free port by spinning up a short-lived socket.
This is not completely correct since some other service could grab the port
during the time when the socket dies and we actually start the server. This diff
makes it possible for the server to accept 0 as the specified port, where it
gets a random port for itself from the system. The new port is then correctly
passed to the frontend/minions.

Test Plan: CI

Reviewed By: ruibm

fbshipit-source-id: bf3c954