Allow multiple root targets to be specified in rulekey differ
Summary: This just allows us to parse a single file once, and check the graph for multiple targets. Since parsing is the highest cost part of the check, it's easiest to handle it in the parser and differ

Test Plan: Updated unit tests. Ran locally with nondeterministic BUCK files and it found differences between multiple unrelated targets

Reviewed By: styurin, ttsugriy

fbshipit-source-id: 284d184