Import fix for namespaced packages
Summary: This diff is basically just an import of https://github.com/pantsbuild/pex/pull/338, but without the tests (we don't have the pex tests in this repo?). This solves a problem where the pex would fail to run any time you had a namespaced package in your path.

Test Plan: eyes

Reviewed By: sdwilsh

fbshipit-source-id: 8d500c3