Omnibus: always link excluded roots dynamically
Summary:
Previously, an excluded root was omitted from the final shared library list if
it's preferred linkage was static, which was inconsistent with non-merged native
link strategies (e.g. if a Python library depended directly onto a "static" C/C++
library, we'd build a shared library for it in the `separate` native link strategy, but
drop it in the `merged` linked strategy).

Test Plan: added test

Reviewed By: styurin

fbshipit-source-id: 6f03b3e