Improve cell-name lookup error messages
Summary:
CellPathResolver.getCellPath used to throw HumanReadableExceptions when
resolution fails. This is probably not the best thing to do for code that's
this low level, as not all calls are with user input cell names, and the actual
place where user input was being handled didn't know to handle this failure.

Instead, return Optional and force callers to handle the case when cell name
has no known mapping. In BuildTargetParser, have the handler throw
BuildTargetParseException, which results in a nicer error message with more
context.

Test Plan:
Say I have a target //some/lib:foo which references a target with an invalid
cell name.

Old behavior:

  buck targets //some/lib:foo
  BUILD FAILED: In cell rooted at /repo_root: cell named 'plaj' is not defined.

New behavior:

  buck targets //some/lib:foo
  BUILD FAILED: //some/lib:foo: parameter 'deps': Unable to find the target plaj//other:foo.
  'plaj//other:foo' references an unknown cell.

Reviewed By: ttsugriy

fbshipit-source-id: 6316ded