rust: add -Cmetadata to distinguish distinct crates with the same name
Summary:
Rust has a flat namespace for crates, but has a mechanism to distinguish
distinct crates with the same name via the `-Cmetadata=xxx` flag.

Without this, the compiler will complain if a single executable tries
to use two crates with the same name, as their symbols may clash. But
`-Cmetadata` causes their symbols to be mangled to guarantee they're
distinct.

This uses the same scheme as cargo - ie, including a truncated md5 hash.
Unlike cargo, this is the hash of the full target name, rather than version
info.

Test Plan: added unit test

Reviewed By: bolinfest

fbshipit-source-id: 93714a3