Pass absolute paths to symlinks in SRCS to genrules
Summary:
`SRCS` environment variable contains a list of absolute paths to inputs specified in `srcs` parameter of a genrule. These paths are outside of sandbox environment and it's blocking proper support of sandboxing.

The solution is to replace it with a list of absolute paths to symlinks in genrule's srcs directory. This directory is created during genrule execution and resembles the directory structure of inputs specified in srcs.

Test Plan: Integration test

Reviewed By: ttsugriy

fbshipit-source-id: eb32855