JS Rules: change file / library artefact extension
Summary:
Changes the extensions of build outputs of `JsLibrary` and `JsFile` from `.json` to `.jslib` and `.jsfile`.

The output data format of these rules is an implementation detail of the worker, and might not be JSON. Using different extensions also makes identifying files in `buck-out` a bit easier.

Test Plan: adapted integration tests

Reviewed By: dinhviethoa

fbshipit-source-id: 1aea56f