Extract a bunch of logic from NdkCxxPlatforms
Summary:
A large part of NdkCxxPlatformTargetConfiguration is really just derived
from the TargetCpuType, but in a rather hard-to-follow way. This
extracts all of that logic out into the TargetCpuType enum values
themselves.

AbstractNdkCxxPlatformTargetConfiguration is then updated to just return
the values from the underlying cpu type (follow-ups will likely remove
that indirection).

Test Plan: CI

Reviewed By: styurin

fbshipit-source-id: a83aa40