Switch SQLite synchronous to normal (from off)
Summary:
We see occasional DB corruption, which tends to result in people asking
for help.  SQLite is supposed to be quite robust, but turning SYNCHRONOUS OFF is
one of the known ways to risk corruption.  NORMAL seems to perform similarly
in my local tests, so let's give that a try.

Test Plan: local builds

Reviewed By: dinhviethoa

fbshipit-source-id: fbeaa8c