Suppress errors more thoroughly
Summary:
`javac`'s `Log` class keeps track of the locations at which diagnostics have been emitted, in an attempt
to avoid issuing the same error twice. When we suppress an error, we should remove it from this list
so that we can report our own errors at that location.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 7845a45