Fix crash on member type that doesn't exist
Summary:
When code referred to a member type that didn't exist (whose parent type was available on the
classpath), source-only ABI generation was crashing. Make it a more graceful error.

Test Plan:
Add an import for a member type that doesn't exist, making sure that the parent type is required
for source only ABI. Reference that type somewhere in the interface of your class. Build. Observe
nice `cannot find symbol` error.

Reviewed By: dreiss

fbshipit-source-id: a2fc2d8