ActionGraphCache with > 1 capacity
Summary: Currently, the ActionGraphCache can only store the most recent actiongraph result. This leads to a lot of misses if the user switches between a small number of build commands frequently. This diff makes ActionGraphCache into a LRU cache with a configurable capacity. The capacity will default to one, but it is configurable in `.buckconfig`.

Test Plan: CI and new method in ActionGraphCacheTest

Reviewed By: ttsugriy

fbshipit-source-id: 1fd0702