Expose module paths as outputs from PreDexMerge
Summary:
`PreDexMerge` is the source of knowledge for where dex outputs go. In order to expose modular dexes to exopackage
we need to expose an accessor for the output paths.

This diff is part of a stack which enables modules to be hotswapped during exo-installs.
This stack includes diffs to enable the following workflow:
* Assemble modular dexes (via `PreDexMerge`)
* Locate modular dexes (via extracting info from the `PreDexMerge` and feeding the output into `ExopackageInfo`)
* Install those modular dexes to a new exopackage location: `/data/local/tmp/modular-dex/` (via `ExopackageInstaller`)
* Load the modular-dex location separately into a new classloader (via `DelegatingClassloader`) that can be thrown away and rebuilt on-demand so that the code can be hotswapped
* Enable the app to register a callback and be notified of/respond to the availability of new code definitions (via `ExoHelper`)

Test Plan: Integration and E2E tests further up the stack

Reviewed By: dreiss

fbshipit-source-id: 658c0f8