Pipe module exopackage info through AndroidBinaryFilesInfo
Summary:
`AndroidBinaryFilesInfo` now populates `ExopackageInfo` with the list of modular dex files and their metadata, by querying `PreDexMerge`. This info will later be fed to `ExopackageInstaller` in order to install the modular dex files.

This diff is part of a stack which enables modules to be hotswapped during exo-installs.
This stack includes diffs to enable the following workflow:
* Assemble modular dexes (via `PreDexMerge`)
* Locate modular dexes (via extracting info from the `PreDexMerge` and feeding the output into `ExopackageInfo`)
* Install those modular dexes to a new exopackage location: `/data/local/tmp/modular-dex/` (via `ExopackageInstaller`)
* Load the modular-dex location separately into a new classloader (via `DelegatingClassloader`) that can be thrown away and rebuilt on-demand so that the code can be hotswapped
* Enable the app to register a callback and be notified of/respond to the availability of new code definitions (via `ExoHelper`)

Test Plan: Integration and E2E tests further up the stack

fbshipit-source-id: b6d464a