Add DelegatingClassLoader
Summary:
Adds a new classloader which will be enabled for exopackage-for-modules builds. This classloader maintains a "delegate" classloader which loads the dexes from
the `modular-dex` exopackage dir. The delegate can be `reset(withListOfNewJars)` in order to throw away old class definitions and load a new set.

This diff is part of a stack which enables modules to be hotswapped during exo-installs.
This stack includes diffs to enable the following workflow:
* Assemble modular dexes (via `PreDexMerge`)
* Locate modular dexes (via extracting info from the `PreDexMerge` and feeding the output into `ExopackageInfo`)
* Install those modular dexes to a new exopackage location: `/data/local/tmp/exopackage/<pkg>/modular-dex/` (via `ExopackageInstaller`)
* Load the modular-dex location separately into a new classloader (via `DelegatingClassloader`) that can be thrown away and rebuilt on-demand so that the code can be hotswapped
* Enable the app to register a callback and be notified of/respond to the availability of new code definitions (via `ExoHelper`)

Test Plan:
Integration/E2E tests lower in the stack, and manual tests in my demo app, the output of which can be seen here:

https://pxl.cl/9t3K

Reviewed By: dreiss

fbshipit-source-id: e950585