Add OnModulesChangedCallback and registration
Summary:
We need to provide apps with a way to respond to hotswap events. Here we introduce an `ExoHelper` class which provides a registry of callbacks which will be notified of hotswap events. This will be hooked up to actual events in a diff later in the stack.

This diff is part of a stack which enables modules to be hotswapped during exo-installs.
This stack includes diffs to enable the following workflow:
* Assemble modular dexes (via `PreDexMerge`)
* Locate modular dexes (via extracting info from the `PreDexMerge` and feeding the output into `ExopackageInfo`)
* Install those modular dexes to a new exopackage location: `/data/local/tmp/exopackage/<pkg>/modular-dex/` (via `ExopackageInstaller`)
* Load the modular-dex location separately into a new classloader (via `DelegatingClassloader`) that can be thrown away and rebuilt on-demand so that the code can be hotswapped
* Enable the app to register a callback and be notified of/respond to the availability of new code definitions (via `ExoHelper`)

Test Plan:
Tested in my demo app (will eventually be checked in somewhere, in the buck repo as an example perhaps). For now, here's a demo video showing a hotswap installation and you can see the app react to the completion of the installation:

https://pxl.cl/9t3K

Reviewed By: dreiss

fbshipit-source-id: 51c1128