Add a BroadcastReceiver to automatically trigger modular dex refresh
Summary:
This diff creates a `BroadcastReceiver` that will be triggered by buck when an exo-install has completed. It can then trigger a reload of the modular dex files as well as any user-registered callbacks

This diff is part of a stack which enables modules to be hotswapped during exo-installs.
This stack includes diffs to enable the following workflow:
* Assemble modular dexes (via `PreDexMerge`)
* Locate modular dexes (via extracting info from the `PreDexMerge` and feeding the output into `ExopackageInfo`)
* Install those modular dexes to a new exopackage location: `/data/local/tmp/exopackage/<pkg>/modular-dex/` (via `ExopackageInstaller`)
* Load the modular-dex location separately into a new classloader (via `DelegatingClassloader`) that can be thrown away and rebuilt on-demand so that the code can be hotswapped
* Enable the app to register a callback and be notified of/respond to the availability of new code definitions (via `ExoHelper`)

Test Plan: I can send the broadcast and see the refresh in the demo app.

Reviewed By: dreiss

fbshipit-source-id: 09ee8fc