Add a helper function for creating an Intent
Summary: Apps which want to take advantage of hotswapping need to do dynamic class lookups in order to trigger new classloads of hotswapped code. This provides a helper function for creating an `Intent` to launch a component from a `String` rather than a `Class` instance.

Test Plan: Manually tested in my demo app

Reviewed By: dreiss

fbshipit-source-id: 3ff9c60