Add a helper function to refresh a given Fragment in place
Summary: When we hotswap a `Fragment` definition, we want to be able to easily trigger a `FragmentTransaction` to replace the `Fragment` currently on-screen with a new instance constructed from the newly hotswapped class definition. This provides a helper method to do that, maintaining original construction args and optionally saving and restoring any saved instance state.

Test Plan: Tested manually in my demo app

Reviewed By: dreiss

fbshipit-source-id: ac2c1d1