Add a helper function to restart the app
Summary: When we hotswap new code into the app, sometimes we don't have a choice but to restart the app in order to reset state. This provides a convenient helper method to do so by setting a `PendingIntent` and then killing the current process.

Test Plan: Tested manually in my demo app.

Reviewed By: dreiss

fbshipit-source-id: 9db99ee