Add compilation stub for support-v4 and support fragment helper functions.
Summary:
We need to provide helper methods for `support.v4.Fragment` but we don't want to depend on a specific instance of the support lib (or one of the sliced versions) and thus force a dependency on our users. This diff contains a minimal shim of the support fragment interface, just enough for us to compile our helper functions.
The helper functions for instantiating and refreshing fragments are reproduced for support fragments.

Test Plan: Tested in my demo app

Reviewed By: dreiss

fbshipit-source-id: bdfe7dd