Reduce scope of s/c/f/b/rules:build_rule
Summary:
s/c/f/b/rules has two large targets, :build_rule and :rules. One of the
main reasons for :build_rule is that some subpackages of c.f.b.rules
depend on the BuildRule class. :rules then depends on those subpackages.

Some consumers of the c.f.b.rules package depend only on :build_rule and
not on :rules. When a consumer does this, it causes us to tend to put things
in :build_rule. This limits our ability to modify the c.f.b.rules package.

This moves a large part of rules:build_rule into rules:rules and updates
a handful of consumers to just depend on rules:rules. As the scope of
:build_rule increases it makes it really hard to add things to
:build_rule and its direct dependents that needs things in :rules. By
limiting the scope of :build_rule and changing a bunch of
non-c.f.b.rules consumers to just consume :rules, it becomes easier to
make changes.

IMO, we should limit the visibility of some of these targets as most
consumers should probably just consume :rules (or some other designated
target exposing the public api), but that isn't in this change.

Test Plan: `buck build buck test/...`, CI

Reviewed By: styurin

fbshipit-source-id: ec4a472