Create only a single executor service for AdbHelper
Summary:
For concurrent install we have multiple calls into AdbHelper. This
changes AdbHelper to create just a single executor service and share that
across multiple calls. The service is closed when the AdbHelper is closed. The
AdbHelper is closed when the execution context is closed.

The main benefit right now of this is that all device interaction happens on the
same thread (or threads if >1 device) and so is easier to follow in traces. In the
future, I think it'll probably make sense to take this a step further and just have
one dedicated single threaded executor service for each device.

Test Plan:
CI, `buck install -c install.concurrent_install=enabled foo` see
that all calls are on same thread.

Reviewed By: asp2insp

fbshipit-source-id: 58820de