rust: add experimental support for procedual macros
Summary:
The Rust compiler rustc has a stable API to allow procedual macros
- ie, a library crate can be compiled as a proc-macro crate type.
When such a crate is used as an input dependency, rather than being
used as a source of definitions for the current crate, it is
dynamically loaded by the compiler and can provide various programatic
source transformations.

We can handle this as an extension to the current rust_library rule
with the boolean "proc_macro" flag.  When set, the crate is compiled
with --crate-type=proc-macro, and is named as a normal shared object
for the platform would be.

TODO: The crate is currently compiled with the same CxxPlatform as
the final target crates are, rather than the CxxPlatform used to
compile rustc itself. This works for non-cross-builds (where the
CPU arch is the same and the rest of the platform is close-enough),
but it should have an override to set the CxxPlatform used to build
the toolchain to guarantee compatibility.

Test Plan: Unit test added.

Reviewed By: andrewjcg

fbshipit-source-id: 96ec605