Use a prettier name for rulekey keys from immutables
Summary:
Currently, if a Value.Immutable object adds one of its values to its
rulekey (w/ AddToRuleKey on the method) we use the method name as the
key. This is ugly as it includes the get/is prefix. When we manually add
such values to rulekeys we use the prettier version of the name with
get/is removed. This updates ValueMethodValueExtractor to use the
prettier name by stripping the prefix and lowercasing the following
letter.

Test Plan: CI

Reviewed By: styurin

fbshipit-source-id: 808ee1b