Test that Buck main jar and module jars have no duplicate classes
Summary:
It is possible to have the same classes in both the main Buck binary and module jars. This situation would result in having the same class to be loaded by different class loaders which means these classes are no longer equal. This a path to some nasty failures.

This change adds a test that analyzes the Buck main binary (`buck_server`) and module jars and reports any classes that are in multiple jars.

Reviewed By: ttsugriy

fbshipit-source-id: 5c00813