Run local build in parallel with distributed build
Summary:
Rather than waiting until the end of a distributed build to kick off a local build, the local build is now started at the same time as the distributed build.

Synchronization has been added to ensure that the local build does not attempt to fetch build rules from the cache until they have finished building remotely.

This will reduce the overall build time as it means that cache items are fetched immediately when they are available (rather than in a big batch at the end of the build), and also because uncachable steps will start building immediately once dependencies are available.

Test Plan: ci

Reviewed By: shivanker

fbshipit-source-id: 2f80cd1