Make prebuilt_python_library work with inplace python binaries
Summary: Right now, .whl files do not work with inplace binaries. This is because we didn't map prebuiltDeps into anything when constructing the link tree. This just links all of the prebuilt library files into a new subdirectory that can be included by the .pex shim.

Test Plan: Tested with a simple project that both -c python.package_style=inplace|standalone worked with a .whl file as a prebuilt_python_library. Also added an integration test that iterates through the different configurations and makes sure that at runtime we can import files from a .whl file.

Reviewed By: ttsugriy

fbshipit-source-id: a25bd08