Add query_targets_and_outputs macro
Summary:
There are some projects that wish to have access to both the target name, and the eventual output of that target as command line paramters in genrules. This allows one to do that for rules that have output (It's effectively smashing query_targets and query_outputs into one).

I was going to make something more generic to allow you to query arbitrary attributes, but that requires a bit more threading of the parser state than I wanted to do just now.

Test Plan: Added new tests, old ones pass, verified that docs look right.

Reviewed By: ttsugriy

fbshipit-source-id: f8757f1