Allow specifying Buck module information in an annotation
Summary:
This change modifies `BuckModule` to be an annotation that contains information about a Buck module instead of being a plugin. It hides the implementation details of a Buck module in generated code and make the process of adding new modules much easier.

Summary of changes:

- `BuckModule` is converted to an annotation that can provide an ID of a module and it's dependencies (not used at this moment)
- Added an annotation processor that generates an adapter plugin for pf4j framework and `MANIFEST.MF` file with information about the plugin.
- Changed ant build and Buck build to use the new annotation processor when requested.
- `ZipRulesModule` is converted to use this approach.

Test Plan: Try to build a `zip_rule` target.

Reviewed By: ttsugriy

fbshipit-source-id: f925e51