Make HashedFileTool hold a SourcePath
Summary:
HashedFileTool uses a hack to add the path to a file directly to
rulekeys.

This instead updates it from using a Path directly to using a
PathSourcePath.

Outside of tests, these are almost exclusively created from BuckConfig
and various views of that, and so it's straightforward for users to
create PathSourcePath instead of Path.

Test Plan: CI

Reviewed By: styurin

fbshipit-source-id: aaa808e