Add AbstractTool
Summary:
This adds a simple AbstractTool that provides a default implementation
of getDeps() that uses BuildableSupport.deriveDeps. This can't be put in
Tool directly because Tool can't depend on BuildableSupport.

Updates all Tool implementations to derive from AbstractTool.

All the implementations of getDeps() that are just derived from the
fields added to the rulekey are deleted.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: ae18b7f