Make ExternalJavac a little easier to understand
Summary:
ExternalJavac has a bunch of weird behavior based on if it's using an
external Path, an internal PathSourcePath or an internal
BuildTargetSourcePath.

This changes things around a bit and adds delegation to a separate
underlying Tool to make things easier to understand.

This allows us to remove ExternalJavac's custom appendToRuleKey() and
getDeps() functions.

Test Plan: CI

Reviewed By: jkeljo

fbshipit-source-id: 64f199b